Ported over the power meter gui and interface.
Added over the files for the process_control_interface, process_interface, and simple_data_interface
These are abstract classes that implement simple interfaces to data. They were also updated with the new
interface implementation protocol.

The powermetergui was updated with the appropriate qudi modifications
the TLPMpowermeter was also updated with the correct imports.

The UI file was updated minorly to add tool text to the buttons to make it more usable.