Refactored galvoLogic into hardware and logic modules (#20)
galvoLogic was split up so that the queryLoop and update display logic
are still there. However, the actual interface to hardware was moved to
galvo_intefuse. This was done because we want to use the galvo as a
virtual piece of hardware in the scanner.

Galvo gui was demonstrated to fully work with the position and voltages.